campaña de marketin completa

+++ server/src/main/java/com/vocaltech/api/domain/campaigns/CampaignEmail.java
@@ -33,6 +33,6 @@ public class CampaignEmail {
     private boolean hasAttachments; // Si el email lleva archivos adjuntos
 
     @Column(nullable = false)
-    private int delayDays; // Cuántos días después del primer email se envía
+    private double delayDays; // Cuántos días después del primer email se envía
 }
